Update example naming and split into different files for table provider and table function
