+++ examples/datafusion-ffi-example/Cargo.toml
@@ -32,5 +32,5 @@ arrow-schema = { version = "55" }
 pyo3-build-config = "0.23"
 
 [lib]
-name = "ffi_table_provider"
+name = "datafusion_ffi_example"
 crate-type = ["cdylib", "rlib"]
